Profileのコントローラー名を変更
