+++ routes/web.php
@@ -20,9 +20,9 @@ Auth::routes();
 Route::post('follow/{user}','FollowsController@store');
 
 //Profile用ルート
-Route::get('profile/{user}','ProfilesController@index');
-Route::get('profile/{user}/edit','ProfilesController@edit');
-Route::patch('profile/{user}','ProfilesController@update');
+Route::get('profile/{user}','ProfileController@index');
+Route::get('profile/{user}/edit','ProfileController@edit');
+Route::patch('profile/{user}','ProfileController@update');
 
 //Post用ルート
 Route::resource('post','PostController');
